refactor SimpleStateRequestHandler to support both keyed state and operator state
