+++ flink-python/src/main/java/org/apache/flink/streaming/api/operators/python/PythonCoProcessOperator.java
@@ -87,6 +87,7 @@ public class PythonCoProcessOperator<IN1, IN2, OUT>
                         config.get(MAP_STATE_WRITE_CACHE_SIZE)),
                 getFlinkMetricContainer(),
                 null,
+                getOperatorStateBackend(),
                 null,
                 null,
                 null,
