+++ flink-python/src/main/java/org/apache/flink/streaming/api/runners/python/beam/BeamPythonFunctionRunner.java
@@ -34,7 +34,9 @@ import org.apache.flink.python.metric.FlinkMetricContainer;
 import org.apache.flink.runtime.memory.MemoryManager;
 import org.apache.flink.runtime.memory.OpaqueMemoryResource;
 import org.apache.flink.runtime.state.KeyedStateBackend;
+import org.apache.flink.runtime.state.OperatorStateBackend;
 import org.apache.flink.streaming.api.operators.python.timer.TimerRegistration;
+import org.apache.flink.streaming.api.runners.python.beam.state.BeamStateRequestHandler;
 import org.apache.flink.util.Preconditions;
 import org.apache.flink.util.function.LongFunctionWithException;
 
@@ -120,6 +122,8 @@ public abstract class BeamPythonFunctionRunner implements PythonFunctionRunner {
 
     @Nullable private final KeyedStateBackend<?> keyedStateBackend;
 
+    @Nullable private final OperatorStateBackend operatorStateBackend;
+
     @Nullable private final TypeSerializer<?> keySerializer;
 
     @Nullable private final TypeSerializer<?> namespaceSerializer;
@@ -187,6 +191,7 @@ public abstract class BeamPythonFunctionRunner implements PythonFunctionRunner {
             ProcessPythonEnvironmentManager environmentManager,
             @Nullable FlinkMetricContainer flinkMetricContainer,
             @Nullable KeyedStateBackend<?> keyedStateBackend,
+            @Nullable OperatorStateBackend operatorStateBackend,
             @Nullable TypeSerializer<?> keySerializer,
             @Nullable TypeSerializer<?> namespaceSerializer,
             @Nullable TimerRegistration timerRegistration,
@@ -199,6 +204,7 @@ public abstract class BeamPythonFunctionRunner implements PythonFunctionRunner {
         this.environmentManager = Preconditions.checkNotNull(environmentManager);
         this.flinkMetricContainer = flinkMetricContainer;
         this.keyedStateBackend = keyedStateBackend;
+        this.operatorStateBackend = operatorStateBackend;
         this.keySerializer = keySerializer;
         this.namespaceSerializer = namespaceSerializer;
         this.timerRegistration = timerRegistration;
@@ -219,7 +225,11 @@ public abstract class BeamPythonFunctionRunner implements PythonFunctionRunner {
 
         stateRequestHandler =
                 getStateRequestHandler(
-                        keyedStateBackend, keySerializer, namespaceSerializer, config);
+                        keyedStateBackend,
+                        operatorStateBackend,
+                        keySerializer,
+                        namespaceSerializer,
+                        config);
 
         // The creation of stageBundleFactory depends on the initialized environment manager.
         environmentManager.open();
@@ -641,16 +651,16 @@ public abstract class BeamPythonFunctionRunner implements PythonFunctionRunner {
     }
 
     private static StateRequestHandler getStateRequestHandler(
-            KeyedStateBackend<?> keyedStateBackend,
-            TypeSerializer<?> keySerializer,
-            TypeSerializer<?> namespaceSerializer,
+            @Nullable KeyedStateBackend<?> keyedStateBackend,
+            @Nullable OperatorStateBackend operatorStateBackend,
+            @Nullable TypeSerializer<?> keySerializer,
+            @Nullable TypeSerializer<?> namespaceSerializer,
             ReadableConfig config) {
-        if (keyedStateBackend == null) {
-            return StateRequestHandler.unsupported();
-        } else {
-            assert keySerializer != null;
-            return new SimpleStateRequestHandler(
-                    keyedStateBackend, keySerializer, namespaceSerializer, config);
-        }
+        return BeamStateRequestHandler.of(
+                keyedStateBackend,
+                operatorStateBackend,
+                keySerializer,
+                namespaceSerializer,
+                config);
     }
 }
