+++ flink-python/src/main/java/org/apache/flink/table/runtime/runners/python/beam/BeamTablePythonFunctionRunner.java
@@ -69,6 +69,7 @@ public class BeamTablePythonFunctionRunner extends BeamPythonFunctionRunner {
                 environmentManager,
                 flinkMetricContainer,
                 keyedStateBackend,
+                null,
                 keySerializer,
                 namespaceSerializer,
                 null,
